+++ providers/eztv.js
@@ -1,6 +1,7 @@
 const async = require("async-q"),
   eztv = require("../lib/eztv"),
   config = require("../config"),
+  colors = require("colors/safe"),
   util = require("../util");
 let helper;
 
@@ -33,7 +34,7 @@ const EZTV = (_name) => {
     search: function*() {
       console.log(name + ": Starting scraping...");
       const eztvShows = yield eztv.getAllShows();
-      console.log(name + ": Found " + eztvShows.length + " shows.");
+      console.log(name + ": Found " + (config.colorOutput ? colors.cyan(eztvShows.length) : eztvShows.length) + " shows.");
       return yield async.mapLimit(eztvShows, config.maxWebRequest, (eztvShow) => {
         return util.spawn(getShow(eztvShow)).catch((err) => {
           util.onError(err);