+++ CHANGELOG.md
@@ -1,13 +1,34 @@
+1.0.3 - It's getting colorful! - 9 March 2016
+==================================================
+
+Features:
+ - Get multiple shows by id at once
+ - Choose to scrape KAT.cr, EZTV.ag or both from config
+ - Color console output (Enable/Disable in config.js)
+
+1.0.2 - It's time to split. - 9 March 2016
+==================================================
+
+Info:
+ - Forked from: https://github.com/popcorn-official/popcorn-api
+ 
+Bug fixes:
+ - Retry on KAT.cr timeout
+ - Updated start command in readme
+
 1.0.1 - What's trening? - 6 March 2016
 ======================================
+
+Info:
+ - Merged pull request: https://github.com/popcorn-official/popcorn-api/pull/1
 
 Bug fixes:
  - Sort by trending
 
-1.0.0 - Let's kick some ass! -  1 March 2016
-============================================
+1.0.0 - Let's kick some ass! - 1 March 2016
+===========================================
 
 Features:
  - Scraping EZTV.ag just like the old API
- - Scraping kat.cr with 17 different providers
+ - Scraping KAT.cr with 17 different providers
  - Able to add more providers for kat.cr scraping